cfg: introduce 'deprecated option' feature
Some options in vshard are going to be eventually deprecated. For
instance, 'weights' will be renamed, 'collect_lua_garbage' may be
deleted since it appears not to be so useful, 'sync_timeout' is
totally unnecessary since any 'sync' can take a timeout per-call.

But the patch is motivated by 'collect_bucket_garbage_interval'
which is going to become unused in the new GC algorithm.

New GC will be reactive instead of proactive. Instead of periodic
polling of _bucket space it will react on needed events
immediately. This will make the 'collect interval' unused.

The option will be deprecated and eventually in some far future
release its usage will lead to an error.

Needed for #147