gc: introduce reactive garbage collector
Garbage collector is a fiber on a master node which deletes
GARBAGE and SENT buckets along with their data.

It was proactive. It used to wakeup with a constant period to
find and delete the needed buckets.

But this won't work with the future feature called 'map-reduce'.
Map-reduce as a preparation stage will need to ensure that all
buckets on a storage are readable and writable. With the current
GC algorithm if a bucket is sent, it won't be deleted for the next
5 seconds by default. During this time all new map-reduce requests
can't execute.

This is not acceptable. As well as too frequent wakeup of GC fiber
because it would waste TX thread time.

The patch makes GC fiber wakeup not by a timeout but by events
happening with _bucket space. GC fiber sleeps on a condition
variable which is signaled when _bucket is changed.

Once GC sees work to do, it won't sleep until it is done. It will
only yield.

This makes GC delete SENT and GARBAGE buckets as soon as possible
reducing the waiting time for the incoming map-reduce requests.

Needed for #147

@TarantoolBot document
Title: Vshard: deprecate cfg option 'collect_bucket_garbage_interval'
It was used to specify the interval between bucket garbage
collection steps. It was needed because garbage collection in
vshard was proactive. It didn't react to newly appeared garbage
buckets immediately.

Since now (0.1.17) garbage collection became reactive. It starts
working with garbage buckets immediately as they appear. And
sleeps rest of the time. The option is not used now and does not
affect behaviour of anything.

I supposed it can be deleted from the documentation. Or left with
a big label 'deprecated' + the explanation above.

An attempt to use the option does not cause an error, but logs a
warning.