introduce CI based on github actions
Currently it's a common practise to use Github Actions for
Tarantool projects. Also seems Travis-CI doesn't work in some
cases.
This patch introduces CI based on Github Actions. Currently it
allows to run tests only on Linux and also pushes rockspec to each
commit to the master branch and for each tag.