storage: introduce bucket_are_all_rw()
In the future map-reduce code it will be needed to be able to
check if all buckets on the storage are in writable state. If they
are - any request can do anything with all the data on the
storage.

Such 'all writable' state will be pinned by a new module
'storage_ref' so as map-reduce requests could execute without
being afraid of the rebalancer.

The patch adds a function bucket_are_all_rw() which is registered
in registry.storage.

The function is not trivial because tries to cache the returned
value. It makes a lot of sense, because the value changes super
rare and the calculation costs a lot (4 lookups in an index by a
string key via FFI + each lookup returns a tuple which is +1 Lua
GC object).

The function is going to be used almost on each map-reduce
request, so it must be fast.

Part of #147