test: fix parallel.test.lua and sync.test.lua
With the current version of test-run, that vshard uses now (d171b6e),
all tests work fine and everything is ok. The problem occures when
attempting to use a later version of test-run starting since 53389a8.
In this case, the parallel.test.lua and sync.test.lua tests start to
fail because the new test-run version catches some exception from tests
and exits with a failure.

This patch should be concidered as a preparation for updating test-run
version for vshard.