ci: add workflow to run against tarantool build
The idea of this workflow is to run against a tarantool build artifact
to verify the integration of the vshard module with an arbitrary
tarantool version. This workflow is not triggered on a push to the repo
but can be run manually since it has the 'workflow_dispatch' trigger.
Moreover, this workflow will be included in the tarantool development
cycle and triggered by the corresponding workflow from the tarantool
project on a push to the master and release branches for verifying
integration of tarantool with vshard.

Part of https://github.com/tarantool/tarantool/issues/4972
Part of https://github.com/tarantool/tarantool/issues/5265
Part of https://github.com/tarantool/tarantool/issues/6056