ci: add reusable testing workflow
The idea of this workflow is to be a pluggable part of some other
workflow from any repo to verify the integration of the vshard module
with an arbitrary tarantool version.

This workflow is not triggered on a push to the repo and cannot be run
manually since it has only the 'workflow_call' trigger. This workflow
will be included in the tarantool development cycle and called by the
corresponding workflow from the tarantool project on a push to the
master and release branches for verifying integration of tarantool with
vshard.

Part of https://github.com/tarantool/tarantool/issues/4972
Part of https://github.com/tarantool/tarantool/issues/5265
Part of https://github.com/tarantool/tarantool/issues/6056