ci: disable fail-fast behaviour for tests
Problem: when tests start and one of them fails, the test run stops, and
other tests are not executed, we exit after the first test failure. It's
not convenient since as a developer I would like to see all test results
and understand the impact of my changes on all project functionalities.