storage: auto enable/disable
While vshard.storage.cfg() is not done, accessing vshard functions
is not safe. It will fail with low level errors like
'access denied' or 'no such function'.

However there can be even worse cases. The user can have universe
access rights. And vshard can be already in global namespace after
require(). So vshard.storage functions are already visible.

The previous patch fixed only the case when function access was
restricted properly. And still fixed it just partially.

New problems are:

- box.cfg{} is already called, but the instance is still
  'loading'. Then data is not fully recovered yet. Accessing is
  not safe from the data consistency perspective.

- vshard.storage.cfg() is not started, or is not finished yet. In
  the end it might be doing something on what the public functions
  depend.

This patch addresses these issues. Now all non-trivial
vshard.storage functions are disabled until vshard.storage.cfg()
is finished and the instance is fully recovered.

They raise an error with a special code. Returning it via
'nil, err' pair wouldn't work. Because firstly, some functions
return a boolean value and are not documented as ever failing.
People would miss this new error.

Second reason - vshard.storage.call() needs to signal the remote
caller that the storage is disabled and it was found before the
user's function was called. If it would be done via 'nil, err',
then the user's function could emulate the storage being disabled.
Or even worse, it could make some changes and then get that error
accidentally by going to another storage remotely which would be
disabled. Hence it is not allowed. Too easy to break something.

It was an option to change vshard.storage.call() signature to
return 'true, retvals...' when user's function was called and
'false, err' when it wasn't, but that would break backward
compatibility. Supporting it only for new routers does not seem
possible.

The patch also drops 'memtx_memory' setting from the config
because an attempt to apply it after calling box.cfg() (for
example, via boot_like_vshard()) raises an error - default memory
is bigger than this setting. It messed the new tests.

Part of #298
Closes #123