Drop Tarantool < 1.10.1 support
The reasons are:

- Rebalancer won't work. It uses fiber.join().

- Bucket discovery on router is extra expensive - it downloads the
  entire _bucket from each storage. That can make the latter and
  the router unresponsive for notable time if bucket count is big
  enough.

- There are other minor hacks which complicate development. For
  instance, box.error.new() couldn't be used.

- Tests don't pass. Thus can't tell what else is broken. Makes no
  sense to support it if it can't be even tested properly

The reason why it is done now is because the code will get more
complicated soon and will use more new features which even 1.10
won't support. This patch should reduce the level of version
dependency complications before they get even worse.

Closes #256
Closes #267

@TarantoolBot document
Title: VShard now supports only Tarantool >= 1.10.1

It used to support 1.9 too, but now it is dropped. To use VShard
> 0.1.19 a user needs Tarantool >= 1.10.1.