Drop periodic Lua GC explicit call
It was introduced because somebody very long time ago complained
that otherwise router goes OOM and the person was sure it was due
to Lua GC working too rare/late.

This of course makes little sense. Either Lua GC was never the
issue, or if it is - then it should be fixed in Lua implementation
in the core. In case anybody still needs that due to any reason,
he can start his own fiber making periodic collectgarbage(). This
has nothing to do with vshard and never had, really.

The patch is done now in scope of general code cleanup from old
unnecessary stuff, like Tarantool < 1.10.0 support.

@TarantoolBot document
Title: VShard cfg option collect_lua_garbage is deprecated

It won't do anything now, only print a warning that it is
deprecated.