util: introduce Tarantool's semver parser
Tarantool's version since recently has semver-like format. It
made impossible to check for features existence if they were
introduced not in the first release of one x.x.x triplet.

An alternative would be to test for feature existence by its
usage, but it is not always possible with ease.

This patch improves version parser in vshard to understand new
version names.

It will be used by a future commit about netbox's return_raw
feature adoption to skip its tests for old versions
(<= 2.10.0-beta2).

Needed for #312