test: run tests in parallel mode via `make test`
Actually, there is no reason not to run tests in parallel mode. Each
test suite has `is_parallel = <True/False>` in the config file which
means whether it is able to be run in parallel or not.

So let's always run tests in parallel and `is_parallel = <True/False>`
will decide whether to do it or not.