cfg: allow uri of any type, not only string
Since 2.10.0-beta2 URI in all APIs can be passed not only as a
string or number but also as a table.

The table can be used to pass options like transport type (plain,
SSL), encryption certificate and key, and potentially more.

VShard always supported only string URIs but now it allows numbers
and tables as well. In the config the replica_object.uri field is
affected by the change.

Part of #325