replicaset: properly compare table URIs
URI comparison is used in order not to recreate netbox connections
on storage/router reconfig. For tables it wasn't working properly
when compared the tables by values. The patch introduces deep
comparison. For example, now {3313} and 3313 are considered the
same URI. If such a change happened in vshard config, it won't
lead to a reconnect.

Closes #325