test: drop meta from vtest.config_new()
The test function used to return 2 values: config and its meta.
The meta could be used to retrieve certain sections of the config
by name, without knowing the automatically generated UUIDs.

The approach appeared to be too clumsy in the next patches. It
would be easier to get UUIDs from the instances like this patch
does.

Part of #325