gc: make it compatible with memtx mvcc
Bucket garbage collector wakes up on _bucket generation change. It
used to happen on each update in _bucket space. When GC fiber
wakes up, it deletes the garbage and goes to sleep until next
generation change.

It didn't work with mvcc because the generation was bumped during
replace before commit, GC fiber woke up, saw no changes, and went
to sleep until next bump. Then _bucket update would commit and
there would be a SENT/GARBAGE bucket visible, but GC fiber
wouldn't even try to find it because the generation didn't change
since last wakeup.

The patch makes so the generation is bumped on commit to _bucket,
not on replace.

Closes #314