test: use prefix 'i' for all instance globals
The instance files storage.lua and router.lua load some modules to
use in _test.lua files in server:exec() functions. It works fine
as long as there aren't many libs and the _test.lua files don't
try to load the same ones.

In some future patches there appeared a problem that certain libs
are needed both in _test.lua files and on the instances as
globals to use them in :exec() calls. It can't be done because
exec requires the function have no upvalues. The function thinks
the libs referenced in it are from _test.lua file, not from the
instance, and treats them as upvalues.

A workaround is to make in exec() function body a hack like
    local mylib = _G.mylib
then _G is loaded on the instance and not treated as an upvalue.

This patch introduces a more generic solution - use 'i' prefix for
all global libraries loaded on the instances ('i' stands for
'instance').

Then in an :exec() call 'ifiber' would reference the
instance-local lib, while outside of exec the _test.lua file can
use just 'fiber' or 'lfiber' or whatever.

Needed for #314