version: recognize 'entrypoint' release type
Apparently, Tarantool core invented its own release type called
'entrypoint'. It is supposed to be less than 'alpha' and wasn't
supported before. Not supported release types are treated as the
greatest, but this one should have been the lowest.

The support is added now because since one of the recent
'entrypoint' releases SSL started working properly in Enterprise.
Next patch will add a test enabled only for that new version.

Needed for #332