storage: change ERRINJ_NO_RECOVERY behaviour
In new tests it will be necessary to pause the recovery again. But
the old implementation of the injection didn't allow to find out
whether the recovery has already got stuck on it. The new one
makes it possible to wait until the recovery is paused by checking
the injection value.

The name is changed to ERRINJ_RECOVERY_PAUSE to reflect the
behaviour better. A similar injection will be introduced for GC.

Needed for #173