storage: handle _bucket changes in commit triggers
Previously _bucket changes only triggered wakeup of various
background fibers. They didn't validate anything nor changed.

While the patch doesn't do much on the field of validation (but it
is a part of the future patches), it makes _bucket updates do some
changes on commit. Without need to do that after an explicit
transaction like

    _bucket:replace(...)
    bucket_ref.ro_lock = true

Now the replace itself on commit tries to do the right things.

This is important because otherwise replicas don't get any updates
to bucket_ref objects. For example, if a bucket was ACTIVE and
became SENT/GARBAGE, and on the replica it had ro refs > 0, then
on the replica it didn't install `ro_lock` to reject new ro refs.
So new read requests could come and would be accepted.

Now also when a bucket is deleted or becomes garbage, its ref
descriptor is deleted. That fixes a similar problem when an
ACTIVE bucket is just deleted completely.

There are also more updates which handle cases impossible during
normal functioning, but still could be done manually or happen as
a bug. For example, it shouldn't be possible that a RECEIVING
bucket has a ref descriptor at all. Yet if it is detected, the
ref is deleted.

The patch intentionally doesn't install rw_lock on SENDING bucket,
because it is supposed to be installed *before* the bucket becomes
SENDING. Checks like that are a subject of a future patch which
will make _bucket on_replace trigger do sanity checks before
commit happens.

The patch is a part of big work on making _bucket updates more
robust and make replicas aware of and agree with them.

Part of #173