vtest: store wait_timeout in a single place
VShard luatests used to define wait_timeout in their test.lua
file. That led to the timeout being duplicated in all the files
and to a bit clumsy code when the timeout needed to be passed to
the instances.

The patch makes so the timeout is stored in a single place - in
vtest. It is a default timeout for existing waiting functions and
will be used in the future patches too.

Needed for #173