storage: make bucket_unref fail critically
It used to fail with WRONG_BUCKET error. That isn't a good idea,
because if the error ever reaches the router, it will retry the
request.

It can't happen now, but in a future patch it will be possible.
Then it could be that ref succeeded, request started execution,
then bucket was deleted, and unref failed. The last step shouldn't
allow the router to retry the request. Especially if it was an RW
request which is not idempotent.

Part of #173