storage: fail a call if couldn't unref the bucket
vshard.storage.call() takes care of bucket referencing to prevent
its move while the call is in progress. It assumed that if a ref
succeeded, then an unref would also work.

But it is not always true. Now it is quite easy to break, because
if a ref is on a replica, then a master could move the bucket even
while it is being accessed on another instance.

The patch makes so vshard.storage.call() fails if unref fails,
signaling that the bucket was probably deleted together with its
ref counter.

In future patches there are going to be more fail-safe measures
like this one. But for some of them will firstly need to fix
bucket GC so as it would consult replicas if they still have refs.

Part of #173