recovery: turn SENDING into SENT, not GARBAGE
Previously SENT and GARBAGE statuses were in practice the same.
Both could be deleted as soon as have no RO refs (except that it
didn't work on replicas).

But soon it is going to change. GARBAGE would be assigned to a
bucket only if it already has no any refs in the entire
replicaset and won't be able to get new ones.

SENT would mean the bucket still can have at least RO refs and
needs validation whether it has them on any replica in its
replicaset.

The patch makes bucket recovery turn SENDING bucket into SENT in
case it is activated in its destination replicaset. The garbage
collector then is trusted to deal with SENT buckets in a special
way. To be done in future patches.

Part of #173