test: port bucket GC test to luatest
Garbage collection is going to be significantly reworked in order
to prevent deletion of SENT buckets which still can have RO refs
on replicas.

It means more GC tests will be needed and the old ones will need
to change. The patch prepares the existing tests to that by
porting them to luatest.

Part of #173