storage: unify index:min() behaviour
At < 2.1.0 index:min(key) worked as

    index:select(key, {limit = 1, iterator='GE'})

At >= 2.1.0 it started working as:

    index:select(key, {limit = 1})

The new behaviour is good and correct. But still vshard needs to
function on 1.10 too. The patch introduces vshard.util.index_min()
which behaves like index:min() at >= 2.1.0. And .index_has()
helper to check if a key exists in an index, which is usually the
purpose of :min().

Follow up tarantool/tarantool#3167
Needed for #173