test: fix flaky rebalancer/stress
The tests stress_add_remove_several_rs and stress_add_remove_rs
try to change cluster topology during rebalancing to check if
the rebalancing eventually ends without errors.

Moreover, during doing that they also generate some RW load via a
router.

The RW load was generated in a fiber which in the end of a test
case is canceled, then the test ensures that all written data is
available for reading.

The problem is that the RW generator fiber had no fiber cancel
checks. It only called `vshard.router.call()` in a loop until it
succeeds. The function doesn't raise exceptions, so if the fiber
was canceled before the `call()` went through, the call was
constantly failing.

What is worse, requests were failing, but they still could get to
the network. They were appended to the netbox's internal buffer,
but the fiber couldn't wait for a response.

Here is how these orphan requests affected the tests. The tests
did the RW load at least 2 times. If a fiber was canceled wrong on
a first load, then on a second load it messed up with the new
requests by keeping retries of the canceled request. That could,
for example, lead to a duplicate PK error in the `test` space.
Then the rebalancer wouldn't be able to finish - duplicate pks in
2 buckets would prevent their storage on the same instance.

Besides, this canceled fibers constantly spamed about that
clogging the logs.

Closes #309