test: fix flaky test on unstable mvcc
In core Tarantool there is a bug in mvcc which sometimes makes
space:count() ~= #space:select(). That works fine now, but an
existing test breaks on that after one of the next commits.
Including an already released Tarantool version 2.10.0. It means
the test must bypass this bug somehow and live with that as long
as vshard supports 2.10.0 at all.

Needed for #173