storage: introduce _call.bucket_test_gc()
There is a bug about master instance's bucket GC fiber not
respecting RO refs on replicas. The idea of a fix - make the
bucket GC procedure consult replicas before marking a bucket as
GARBAGE. Why this way - see motivation in the main commit.

The present commit introduces a helper - service call
bucket_test_gc(). The function takes bucket IDs and returns which
of them are not ok to be GCed. Additionally, it assumes all the
tested buckets are SENT. Because testing the others for being
collectible does not make sense and looks like a bug - it is
returned as an error then.

The function returns "not ok" bids instead of "ok" ones. Because
most of the tested buckets are supposed to be collectible. Sending
"not ok" bids would produce less network traffic. And is easier to
process in a map-reduce way. Indeed, easier to mark all bids which
are not ok on at least one replica, then for each bid do count how
many replicas are ok with it being removed. Is clearly visible in
the main commit later.

Needed for #173